Put NVUE quirks in their own class
